Adding holiday dates for England and Wales
Iv attempted to add dates for public holidays in England and Wales. Iv not found any full documentation but have guessed on the syntax by looking at so it would probably be wise to double check what iv done. It appears that dates should be in the month,day format. Could some also confirm weather or not it matters what order the holidays are listed, Boxing day usually follows Christmas day but in years where the 25th of December is a Sunday then they swap placed (as will happen this year [2016]) so I have placed them the way round which they occur most frequently.
Sourced from: https://www.gov.uk/bank-holidays#england-and-wales